refactor(game-map-model): move and add some docs
