+++ src/main/scala/scatan/model/ApplicationState.scala
@@ -3,6 +3,7 @@ package scatan.model
 import scatan.lib.mvc.Model
 import scatan.model.game.ScatanGame
 import scatan.model.game.config.ScatanPlayer
+import map.GameMap
 
 final case class ApplicationState(game: Option[ScatanGame]) extends Model.State:
   def createGame(gameMap: GameMap, usernames: String*): ApplicationState =
