+++ src/main/scala/scatan/model/map/GameMap.scala
@@ -1,4 +1,4 @@
-package scatan.model
+package scatan.model.map
 
 import scatan.model.components.Terrain
 import scatan.model.map.*
@@ -11,6 +11,9 @@ import scatan.model.map.HexagonInMap.*
   *
   * @param withSeaLayers
   *   number of concentric circles of hexagons the terrain ones.
+  *
+  * @param tileContentStrategy
+  *   strategy to generate the content of the tiles
   */
 final case class GameMap(
     withTerrainLayers: Int = 2,
@@ -32,14 +35,26 @@ final case class GameMap(
         (this.toContent.toSet & that.toContent.toSet).sizeIs == this.toContent.size
       case _ => false
 
+/** A factory to create game maps.
+  */
 object GameMapFactory:
 
+  /** @return
+    *   a fixed game map for layer 2
+    */
   def defaultMap: GameMap =
     GameMap(tileContentStrategy = TileContentStrategyFactory.fixedForLayer2)
 
+  /** @return
+    *   a random game map for layer 2
+    */
   def randomMap: GameMap =
     GameMap(tileContentStrategy = TileContentStrategyFactory.randomForLayer2)
 
-  val strategies: Iterator[TileContentStrategy] = TileContentStrategyFactory.permutationForLayer2.toIterator
+  private val strategies: Iterator[TileContentStrategy] = TileContentStrategyFactory.permutationForLayer2.toIterator
+
+  /** @return
+    *   the next permutation of the game map for layer 2
+    */
   def nextPermutation: GameMap =
     GameMap(tileContentStrategy = strategies.next())
