+++ src/test/scala/scatan/model/GameTest.scala
@@ -11,6 +11,7 @@ import scatan.model.game.config.ScatanActions.*
 import scatan.model.game.config.{ScatanActions, ScatanPhases, ScatanPlayer, ScatanSteps}
 import scatan.model.game.state.ScatanState
 import scatan.model.game.state.ops.EmptySpotsOps.{emptyRoadSpots, emptyStructureSpots}
+import map.GameMap
 
 class GameTest extends BaseTest:
