+++ src/test/scala/scatan/model/game/ScatanRulesTest.scala
@@ -3,7 +3,7 @@ package scatan.model.game
 import org.scalatest.matchers.should.Matchers.shouldBe
 import scatan.BaseTest
 import scatan.lib.game.GameStatus
-import scatan.model.GameMap
+import scatan.model.map.GameMap
 import scatan.model.game.config.{ScatanActions, ScatanPhases, ScatanPlayer, ScatanSteps}
 import scatan.model.game.state.ScatanState
