+++ src/test/scala/scatan/model/map/GameMapFactoryTest.scala
@@ -2,7 +2,6 @@ package scatan.model.map
 
 import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
 import scatan.BaseTest
-import scatan.model.{GameMap, GameMapFactory}
 
 class GameMapFactoryTest extends BaseTest with ScalaCheckPropertyChecks:
