+++ src/test/scala/scatan/model/map/GameMapTest.scala
@@ -2,7 +2,6 @@ package scatan.model.map
 
 import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
 import scatan.BaseTest
-import scatan.model.GameMap
 import scatan.model.components.UnproductiveTerrain.*
 import scatan.model.map.HexagonInMap.layer
